projects/livecam-ai-assitant/pyproject.toml
[project]
name = "livecam-ai-assitant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "elevenlabs>=2.7.1",
    "gradio>=5.38.0",
    "groq>=0.30.0",
    "gtts>=2.5.4",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.8",
    "langgraph>=0.5.3",
    "opencv-python>=4.12.0.88",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "pyhub>=0.0.11",
    "python-dotenv>=1.1.1",
    "speechrecognition>=3.14.3",
]
